Change NavBar to a component, removed from DashboardView
