+++ client/src/App.vue
@@ -1,7 +1,9 @@
 <template>
+<v-app id="inspire">
   <v-main>      
     <router-view></router-view> 
   </v-main>
+</v-app>
 </template>
 
 <script>
